requirements.txt
opencv-python>=4.8.0
python-dotenv>=1.0.0
google-cloud-vision>=3.4.4
openai>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0
matplotlib>=3.7.0 